refactor: clean up dtos and move packages
