--- src/main/java/kitchenpos/application/dto/OrderTableRequest.java
@@ -1,8 +1,0 @@
-package kitchenpos.application.dto;
-
-public class OrderTableRequest {
-    private Long id;
-    private Long tableGroupId;
-    private int numberOfGuests;
-    private boolean empty;
-}
