+++ src/main/java/kitchenpos/application/dto/request/MenuCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.request;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -9,12 +9,13 @@ import kitchenpos.domain.Product;
 
 public class MenuCreateRequest {
 
-    private String name;
-    private BigDecimal price;
-    private Long menuGroupId;
-    private List<MenuProductCreateRequest> productRequests;
+    private final String name;
+    private final BigDecimal price;
+    private final Long menuGroupId;
+    private final List<MenuProductCreateRequest> productRequests;
 
     private MenuCreateRequest() {
+        this(null, null, null, null);
     }
 
     public MenuCreateRequest(
