+++ src/main/java/kitchenpos/application/dto/request/ProductCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.request;
 
 import java.math.BigDecimal;
 import kitchenpos.domain.Product;
