+++ src/main/java/kitchenpos/application/dto/response/MenuResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.response;
 
 import java.math.BigDecimal;
 import java.util.List;
