+++ src/main/java/kitchenpos/application/dto/response/OrderLineItemResponse.java
@@ -1,21 +1,19 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.response;
 
 import kitchenpos.domain.OrderLineItem;
 
 public class OrderLineItemResponse {
 
     private final Long seq;
-//    private final Long orderId;
     private final Long menuId;
     private final long quantity;
 
-    public OrderLineItemResponse() {
+    private OrderLineItemResponse() {
         this(null, null, 0);
     }
 
     public OrderLineItemResponse(Long seq, Long menuId, long quantity) {
         this.seq = seq;
-//        this.orderId = orderId;
         this.menuId = menuId;
         this.quantity = quantity;
     }
@@ -23,7 +21,6 @@ public class OrderLineItemResponse {
     public static OrderLineItemResponse from(OrderLineItem orderLineItem) {
         return new OrderLineItemResponse(
                 orderLineItem.getSeq(),
-//                orderLineItem.getOrderId(),
                 orderLineItem.getMenuId(),
                 orderLineItem.getQuantity()
         );
@@ -32,10 +29,6 @@ public class OrderLineItemResponse {
     public Long getSeq() {
         return seq;
     }
-
-//    public Long getOrderId() {
-//        return orderId;
-//    }
 
     public Long getMenuId() {
         return menuId;
