+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -3,9 +3,9 @@ package kitchenpos.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.application.OrderService;
-import kitchenpos.application.dto.OrderStatusChangeRequest;
-import kitchenpos.application.dto.OrderRequest;
-import kitchenpos.application.dto.OrderResponse;
+import kitchenpos.application.dto.request.OrderStatusChangeRequest;
+import kitchenpos.application.dto.request.OrderRequest;
+import kitchenpos.application.dto.response.OrderResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
