+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -10,9 +10,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import kitchenpos.OrderTableFixtures;
 import kitchenpos.TableGroupFixtures;
-import kitchenpos.application.dto.OrderRequest;
-import kitchenpos.application.dto.OrderResponse;
-import kitchenpos.application.dto.OrderStatusChangeRequest;
+import kitchenpos.application.dto.request.OrderRequest;
+import kitchenpos.application.dto.response.OrderResponse;
+import kitchenpos.application.dto.request.OrderStatusChangeRequest;
 import kitchenpos.dao.OrderRepository;
 import kitchenpos.dao.OrderTableRepository;
 import kitchenpos.dao.TableGroupRepository;
