+++ src/test/java/kitchenpos/ui/MenuGroupRestControllerTest.java
@@ -12,7 +12,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import java.util.List;
 import kitchenpos.application.MenuGroupService;
-import kitchenpos.application.dto.MenuGroupResponse;
+import kitchenpos.application.dto.response.MenuGroupResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
